dj-homeworks/diplom/requirements.txt
Django==2.2.1
django-ckeditor==5.6.1
pillow==5.3.0
